Implement PA-R12-003 (#912): hex literals with top bit set now encode correctly
Bug: `mov rax, 0xFFFFFFFFFFFFFFFD` silently encoded as `48 b8 00 00
00 00 00 00 00 00` (movabs \$0x0) instead of `48 b8 fd ff ff ff ff
ff ff ff`. Root cause: both hex-literal parsers used
`i64::from_str_radix`, which rejects values > i64::MAX
(0x7FFF_FFFF_FFFF_FFFF). The rejection was silently swallowed via
`unwrap_or(0)` or `Err(())`, producing 0 as the parsed value.

## Bug sites

- `crates/paideia-as-elaborator/src/unsafe_walker.rs` — hex/oct/bin
  branches in `extract_integer_from_span` (path taken by
  `mov rax, imm64` in unsafe blocks).
- `crates/paideia-as/src/cmd_build.rs` — `parse_integer_literal`
  (path taken by module-level `let X : u64 = 0xFFFF...`).

## Fix

Both sites now use `u64::from_str_radix` for hex/oct/bin literals
with a bit-preserving `as i64` cast. Decimal parsing (signed and
unsigned) unchanged.

Also replace `-n` with `n.wrapping_neg()` on negative-decimal
reconstruction in `cmd_build.rs` to fix a latent `i64::MIN` panic
(cheap correctness win alongside the primary fix; `-i64::MIN`
would previously panic).

## Runtime impact on paideia-os

R12 handlers (kind_page.pdx, kind_sched.pdx) use:
- INVOKE_DENIED = 0xFFFFFFFFFFFFFFFD (rights failure sentinel)
- INVOKE_UNSUPPORTED = 0xFFFFFFFFFFFFFFFC (unknown op sentinel)
- 0xDEADBEEF00000004..7 stub sentinels

All were silently returning 0 instead of their intended sentinel.
0 happens to be YIELD_OK / SUCCESS in several contexts, so rights
failures were silently succeeding. Post-fix, sentinels encode
correctly.

Higher-half kernel virtual addresses (0xFFFF_8000_..._+) and NX-bit
page-table entries (bit 63 = 0x8000_0000_0000_0000) also affected;
paideia-os hasn't exercised those yet but they're now correct for
future MM work.

## Test coverage

Unit tests in cmd_build.rs:
- parse_hex_top_bit_set: 0xFFF...FD, 0xDEADBEEF00000007,
  0xFFF...FF, 0x7FFF...FF (i64::MAX regression) all correct
- parse_negative_decimal_still_works: -42, i64::MIN (previously panic)

Integration tests build_emit_imm64_top_bit.rs (5 tests):
- imm64_top_bit_fd: mov rax, 0xFFF...FD -> 48 b8 fd ff ff ff ff ff ff ff
- imm64_top_bit_fc: mov rax, 0xFFF...FC -> 48 b8 fc ff ff ff ff ff ff ff
- imm64_deadbeef: mov rax, 0xDEADBEEF00000007 -> 48 b8 07 00 00 00 ef be ad de
- imm64_all_ones: mov rax, 0xFFF...FF -> 48 b8 ff ff ff ff ff ff ff ff
- imm64_i64_max: mov rax, 0x7FFF...FF -> 48 b8 ff ff ff ff ff ff ff 7f (regression)

Snapshot canary paideia_os_m3_829_byte_snapshot unchanged (existing
paideia-os kernel doesn't use top-bit-set literals in the snapshot
scope).

Manual reproducer verified: `mov rax, 0xFFFFFFFFFFFFFFFD` now emits
`movabs \$0xfffffffffffffffd,%rax`.

## Scope

Not fixed (separate follow-ups filed as needed):
- Silent `unwrap_or(0)` in unsafe_walker.rs:569 for genuinely
  malformed literals (should be a diagnostic).
- Underscore separator in hex literals (0x1_0000_0000): stripped in
  cmd_build.rs but not unsafe_walker.rs. Divergent behavior.
- Duplicated integer-literal parsers across three crates.

+++ crates/paideia-as-elaborator/src/unsafe_walker.rs
@@ -1011,24 +1011,24 @@ fn extract_integer_from_span(
         return Some(val);
     }
 
-    // Try parsing as hex
+    // Try parsing as hex (u64-first so top-bit-set values survive)
     if text.starts_with("0x") || text.starts_with("0X") {
-        if let Ok(val) = i64::from_str_radix(&text[2..], 16) {
-            return Some(val);
+        if let Ok(val) = u64::from_str_radix(&text[2..], 16) {
+            return Some(val as i64);
         }
     }
 
     // Try parsing as octal
     if text.starts_with("0o") || text.starts_with("0O") {
-        if let Ok(val) = i64::from_str_radix(&text[2..], 8) {
-            return Some(val);
+        if let Ok(val) = u64::from_str_radix(&text[2..], 8) {
+            return Some(val as i64);
         }
     }
 
     // Try parsing as binary
     if text.starts_with("0b") || text.starts_with("0B") {
-        if let Ok(val) = i64::from_str_radix(&text[2..], 2) {
-            return Some(val);
+        if let Ok(val) = u64::from_str_radix(&text[2..], 2) {
+            return Some(val as i64);
         }
     }
